# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Constancia Backend API",
    "description": "Backend API contract for the Constancia frontend and Discord bot.",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/campaigns/{id}/characters/": {
      "get": {
        "operationId": "listCharacters",
        "summary": "List campaign characters",
        "tags": ["characters"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "discordUserId": {
                            "type": "string"
                          },
                          "discordName": {
                            "type": "string"
                          },
                          "gameName": {
                            "type": "string"
                          },
                          "campaignId": {
                            "type": "string"
                          },
                          "backstory": {
                            "type": "string"
                          },
                          "notes": {
                            "type": "string"
                          },
                          "systemData": {
                            "type": "object",
                            "additionalProperties": true
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "discordUserId",
                          "discordName",
                          "gameName",
                          "campaignId",
                          "backstory",
                          "notes"
                        ]
                      }
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createCharacter",
        "summary": "Create a character",
        "tags": ["characters"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "discordUserId": {
                    "type": "string"
                  },
                  "backstory": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "systemData": {
                    "type": "object",
                    "additionalProperties": true
                  }
                },
                "required": ["name", "discordUserId"]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "discordUserId": {
                          "type": "string"
                        },
                        "discordName": {
                          "type": "string"
                        },
                        "gameName": {
                          "type": "string"
                        },
                        "campaignId": {
                          "type": "string"
                        },
                        "backstory": {
                          "type": "string"
                        },
                        "notes": {
                          "type": "string"
                        },
                        "systemData": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "discordUserId",
                        "discordName",
                        "gameName",
                        "campaignId",
                        "backstory",
                        "notes"
                      ]
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{id}/characters/{charId}": {
      "get": {
        "operationId": "getCharacter",
        "summary": "Get a character",
        "tags": ["characters"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "charId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "discordUserId": {
                          "type": "string"
                        },
                        "discordName": {
                          "type": "string"
                        },
                        "gameName": {
                          "type": "string"
                        },
                        "campaignId": {
                          "type": "string"
                        },
                        "backstory": {
                          "type": "string"
                        },
                        "notes": {
                          "type": "string"
                        },
                        "systemData": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "discordUserId",
                        "discordName",
                        "gameName",
                        "campaignId",
                        "backstory",
                        "notes"
                      ]
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateCharacter",
        "summary": "Update a character",
        "tags": ["characters"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "gameName": {
                    "type": "string"
                  },
                  "backstory": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "systemData": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "charId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "discordUserId": {
                          "type": "string"
                        },
                        "discordName": {
                          "type": "string"
                        },
                        "gameName": {
                          "type": "string"
                        },
                        "campaignId": {
                          "type": "string"
                        },
                        "backstory": {
                          "type": "string"
                        },
                        "notes": {
                          "type": "string"
                        },
                        "systemData": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "discordUserId",
                        "discordName",
                        "gameName",
                        "campaignId",
                        "backstory",
                        "notes"
                      ]
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}